Refactor exception logic of buildings and classrooms
