+++ server/models/database/subject_db_model.py
@@ -1,6 +1,7 @@
 from datetime import datetime
 
 from beanie import Document, Link, Indexed
+from fastapi import HTTPException, status
 from typing import Self, Annotated
 
 from server.models.database.building_db_model import Building
@@ -22,9 +23,21 @@ class Subject(Document):
         keep_nulls = False
 
     @classmethod
-    async def by_id(cls, id: str) -> Self | None:
-        return await cls.get(id)
+    async def by_id(cls, id: str) -> Self:
+        subject = await Subject.by_id(id)
+        if subject is None:
+            raise SubjectNotFound(id)
+        return subject
 
     @classmethod
     async def by_code(cls, code: str) -> Self | None:
-        return await Subject.find_one(cls.code == code)
+        subject = await Subject.find_one(cls.code == code)
+        if subject is None:
+            raise SubjectNotFound(code)
+        return subject
+
+
+class SubjectNotFound(HTTPException):
+    def __init__(self, subject_info: str) -> None:
+        super().__init__(status.HTTP_404_NOT_FOUND,
+                         f"Subject {subject_info} not found")
